% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpr.R
\name{coexpr}
\alias{coexpr}
\title{Identification of gene pairs coexpressed in at least one of two conditions}
\usage{
coexpr(exprs.1, exprs.2, rth = 0.5, qth = 0.1, r.method = c("pearson",
  "spearman")[1], q.method = c("BH", "holm", "hochberg", "hommel",
  "bonferroni", "BY", "fdr")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as genes and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as genes and columns as samples.}

\item{rth}{the cutoff of r; must be within [0,1].}

\item{qth}{the cutoff of q-value; must be within [0,1].}

\item{r.method}{a character string specifying the method to be used to calculate correlation coefficients.}

\item{q.method}{method for adjusting p values.}
}
\value{
a data frame containing gene pairs coexpressed in either condition 1 or condition 2
}
\description{
This function identifies gene pairs coexpressed in either condition 1 or condition 2.
}
\examples{
#rFilter()
}
\keyword{coexpression}
