% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparecor.R
\name{comparecor}
\alias{comparecor}
\title{Compare gene-gene correlation coefficients under two conditions}
\usage{
comparecor(exprs.1, exprs.2, r.method = c("pearson", "spearman")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as genes and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as genes and columns as samples.}

\item{r.method}{a character string specifying the method to be used to calculate correlation coefficients.}
}
\value{
a data frame
}
\description{
This function calculates correlation coefficients of gene pairs in condition 1 and condition 2 and compare them using Fisher's Z-transformation.
}
\examples{
#comparecor()
}
\keyword{coexpression}
