% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprs.1}
\alias{exprs.1}
\title{exprs.1}
\format{A data frame with 200 genes and 14 samples.}
\usage{
exprs.1
}
\description{
expression of 200 genes in 14 samples (GSM94988 to GSM95001) of yeast after pulses
2 g/l glucose, \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4158}.
}
\keyword{datasets}
