% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcoexp.R
\name{rFilter}
\alias{rFilter}
\title{A rFilter Function}
\usage{
rFilter(exprs.1, exprs.2, rth = 0.5, qth = 0.1, r.method = c("pearson",
  "spearman")[1], q.method = c("BH", "holm", "hochberg", "hommel",
  "bonferroni", "BY", "fdr")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as variables (genes) and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as variables (genes) and columns as samples.}

\item{rth}{the cutoff of r; must be within [0,1].}

\item{qth}{the cutoff of q-value; must be within [0,1].}
}
\description{
This function is used to identify coexpressed links (gene pairs) in either condition 1 or condition 2.
}
\examples{
rFilter()
}
\keyword{coexpression}
