% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpr.R
\name{coexpr}
\alias{coexpr}
\title{Identification of gene pairs coexpressed in at least one of two conditions}
\usage{
coexpr(exprs.1, exprs.2, rth = 0.5, qth = 0.1, r.method = c("pearson",
  "spearman")[1], q.method = c("BH", "holm", "hochberg", "hommel",
  "bonferroni", "BY", "fdr")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as genes and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as genes and columns as samples.}

\item{rth}{the cutoff of r; must be within [0,1].}

\item{qth}{the cutoff of q-value; must be within [0,1].}

\item{r.method}{a character string specifying the method to be used to calculate correlation coefficients.}

\item{q.method}{method for adjusting p values.}
}
\value{
a data frame containing gene pairs that are coexpressed in at least one of the conditions with the criteria that the absolute value of at least one of their correlation coefficients (cor.1 and/or cor.2) is greater that rth with q value (q.1 and/or q.2) less than qth. It has the following columns:
  \item{\code{Gene.Pair}}{Gene pair}
  \item{\code{cor.1}}{correlation coefficients in condition 1}
  \item{\code{cor.2}}{correlation coefficients in condition 2}
  \item{\code{p.1}}{p value under null hypothesis that correlation coefficient in condition 1 equals to zero}
  \item{\code{p.2}}{p value under null hypothesis that correlation coefficient in condition 2 equals to zero}
  \item{\code{p.diffcor}}{p value under null hypothesis that difference between two correlation coefficients under two conditions equals to zero using Fisher's r-to-Z transformation}
  \item{\code{q.1}}{adjusted p value under null hypothesis that correlation coefficient in condition 1 equals to zero}
  \item{\code{q.2}}{adjusted p value under null hypothesis that correlation coefficient in condition 2 equals to zero}
  \item{\code{q.diffcor}}{adjusted p value under null hypothesis that the difference between two correlation coefficients under two conditions equals to zero using Fisher's r-to-Z transformation}
}
\description{
This function identifies gene pairs coexpressed in at least one of two conditions.
}
\examples{
data(exprs4158)
allowWGCNAThreads()
res=coexpr(exprs.1 = exprs.1, exprs.2 = exprs.2, r.method = "spearman")
#The result is a data frames.
str(res)
}
\keyword{coexpression}
