% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpr.R
\name{coexpr}
\alias{coexpr}
\title{Identification of gene pairs coexpressed in at least one of two conditions}
\usage{
coexpr(exprs.1, exprs.2, rth = 0.5, qth = 0.1, r.method = c("pearson",
  "spearman")[1], q.method = c("BH", "holm", "hochberg", "hommel",
  "bonferroni", "BY", "fdr")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as genes and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as genes and columns as samples.}

\item{rth}{the cutoff of r; must be within [0,1].}

\item{qth}{the cutoff of q-value; must be within [0,1].}

\item{r.method}{a character string specifying the method to be used to calculate correlation coefficients.}

\item{q.method}{method for adjusting p values.}
}
\value{
a data frame containing gene pairs coexpressed in at least one of the conditions with the criteria that at least one of their correlation coefficients (cor.1 and/or cor.2) is greater that rth with q value (q.1 and/or q.2) less than qth. It has the following columns:
  \item{\code{Gene.1}}{Gene ID}
  \item{\code{Gene.2}}{Gene ID}
  \item{\code{cor.1}}{correlation coefficients in condition 1}
  \item{\code{cor.2}}{correlation coefficients in condition 2}
  \item{\code{p.1}}{p value under null hypothesis that correlation coefficient in condition 1 equals to zero}
  \item{\code{p.2}}{p value under null hypothesis that correlation coefficient in condition 2 equals to zero}
  \item{\code{p.diffcor}}{p value under null hypothesis that difference between two correlation coefficients under two conditions equals to zero using Fisher's r-to-Z transformation}
  \item{\code{q.1}}{adjusted p value under null hypothesis that correlation coefficient in condition 1 equals to zero}
  \item{\code{q.2}}{adjusted p value under null hypothesis that correlation coefficient in condition 2 equals to zero}
  \item{\code{q.diffcor}}{adjusted p value under null hypothesis that the difference between two correlation coefficients under two conditions equals to zero using Fisher's r-to-Z transformation}
  \item{\code{cor.diff}}{difference between correlation coefficients in condition 2 and condition 1}
  \item{\code{type}}{can have value "same signed", "diff signed", or "switched opposites". "same signed" indicates that the gene pair has same signed correlation coefficients under both conditions. "diff signed" indicates that the gene pair has opposite signed correlation coefficients under two conditions and only one of them passed the criteria that the absolute correlation coefficients greater than rth and q value less than qth. "switched opposites" indicates that the gene pair has opposite signed correlation coefficients under two conditions and both of them passed the criteria that the absolute correlation coefficients greater than rth and q value less than qth.}
}
\description{
This function identifies gene pairs coexpressed in either condition 1 or condition 2.
}
\details{
diffcoexp function identifies differentially coexpressed links (DCLs) and differentially coexpressed genes (DCGs). DCLs are gene pairs with significantly different correlation coefficients under two conditions (de la Fuente 2010, Jiang et al., 2016). DCGs are genes with significantly more DCLs than by chance (Yu et al., 2011, Jiang et al., 2016). It takes two gene expression matrices or data frames under two conditions as input, calculates gene-gene correlations under two conditions and compare them with Fisher's Z transformation, filter the correlation with the rth and qth and the correlation changes with r.diffth and q.diffth. It identifies DCGs using binomial probability model (Jiang et al., 2016).

The main steps are as follows:

a). In this step, gene pairs (links) coexpressed in at least one condition are identified using the criteria that at least one of the the correlation coefficients under two conditions having absolute value greater than the threshold rth and the adjusted p value less than the threshold qth. The links passed the criteria are included in CO.links.

b). In this step, differentially coexpressed gene pairs (links) are identified from CO.links using the criteria that the absolute value of the difference between the two correlation coefficients is greater the threshold r.diffth and the adjusted p value is less than the threshold q.diffth. The links passed the criteria are included in DCLs and DC.links.

c). In this step, the DCLs are classified into three categories: "same signed", "diff signed", or "switched opposites". "same signed" indicates that the gene pair has same signed correlation coefficients under both conditions. "diff signed" indicates that the gene pair has opposite signed correlation coefficients under two conditions and only one of them passed the criteria that the absolute correlation coefficients greater than the threshold rth and adjusted p value less than the threshold qth. "switched opposites" indicates that the gene pair has opposite signed correlation coefficients under two conditions and both of them passed the criteria that the absolute correlation coefficients greater than the threshold rth and adjusted p value less than the threshold qth.

d). In this step, all the genes in DCLs are tested for their enrichment, i.e, whether they have more DC.links than by chance using binomial probability model (Jiang et al., 2016)
}
\examples{
#rFilter()
}
\references{
1. de la Fuente A. From “differential expression” to “differential networking” – identification of dysfunctional regulatory networks in diseases. Trends in Genetics. 2010 Jul;26(7):326–33.

2. Jiang Z, Dong X, Li Z-G, He F, Zhang Z. Differential Coexpression Analysis Reveals Extensive Rewiring of Arabidopsis Gene Coexpression in Response to Pseudomonas syringae Infection. Scientific Reports [Internet]. 2016 Dec [cited 2017 Sep 20];6(1). Available from: http://www.nature.com/articles/srep35064

3. Yu H, Liu B-H, Ye Z-Q, Li C, Li Y-X, Li Y-Y. Link-based quantitative methods to identify differentially coexpressed genes and gene pairs. BMC bioinformatics. 2011;12(1):315.
}
\author{
Wenbin Wei
}
\keyword{coexpression}
