% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcoexp.R
\name{diffcoexp}
\alias{diffcoexp}
\title{Differential co-expression analysis}
\usage{
diffcoexp(exprs.1, exprs.2, rth = 0.5, qth = 0.1, r.diffth = 0.5,
  q.diffth = 0.1, q.dcgth = 0.1, r.method = c("pearson", "kendall",
  "spearman")[1], q.method = c("BH", "holm", "hochberg", "hommel",
  "bonferroni", "BY", "fdr")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as genes and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as genes and columns as samples.}

\item{rth}{the cutoff of r; must be within [0,1].}

\item{qth}{the cutoff of q-value (adjusted p value); must be within [0,1].}

\item{r.diffth}{the cutoff of absolute value of the difference between the correlation coefficients of the two conditions; must be within [0,1].}

\item{q.diffth}{the cutoff of q-value (adjusted p value) of the difference between the correlation coefficients of the two conditions; must be within [0,1].}

\item{q.dcgth}{the cutoff of q-value (adjusted p value) of the genes enriched in the differentilly correlated gene pairs between the two conditions; must be within [0,1].}

\item{r.method}{a character string specifying the method to be used to calculate correlation coefficients.}

\item{q.method}{method for adjusting p values.}
}
\value{
a list of two data frames.

The DCGs data frame contains the following columns:
  \item{\code{Gene}}{Gene ID}
  \item{\code{All.links}}{Number of links with the absolute correlation coefficients greater than rth and q value less than qth in at least one condition}
  \item{\code{DC.links}}{Number of links that passed the criteria for All links and the criteria that the absolute differences between the correlation coefficients in the two condition greater than r.diffth and q value less than q.diffth}
  \item{\code{DCL_same}}{Number of subset of DC links with same signed correlation coefficients in both conditions}
  \item{\code{DCL_diff}}{Number of subset of DC links with opposite signed correlation coefficients in two conditions but only one of them with the absolute correlation coefficients greater than rth and q value less than qth}
  \item{\code{DCL_switch}}{Number of subset of DC links with opposite signed correlation coefficients in two conditions and both of them with the absolute correlation coefficients greater than rth and q value less than qth}

The DCLs data frame contains the following columns:
  \item{\code{Gene.1}}{Gene ID}
  \item{\code{Gene.2}}{Gene ID}
  \item{\code{cor.1}}{correlation coefficients in condition 1}
  \item{\code{cor.2}}{correlation coefficients in condition 2}
  \item{\code{p.1}}{p value of correlation coefficients in condition 1}
  \item{\code{p.2}}{p value of correlation coefficients in condition 2}
  \item{\code{p.diffcor}}{p value of the test of significance for the difference between two correlation coefficients under two conditions using Fisher's r-to-Z transformation}
  \item{\code{q.1}}{adjusted p value of correlation coefficients in condition 1}
  \item{\code{q.2}}{adjusted p value of correlation coefficients in condition 2}
  \item{\code{q.diffcor}}{adjusted p value of the test of significance for the difference between two correlation coefficients under two conditions using Fisher's r-to-Z transformation}
  \item{\code{cor.diff}}{difference between correlation coefficients in condition 2 and condition 1}
  \item{\code{type}}{can have value "same signed", "diff signed", or "switched opposites". "same signed" indicates that the gene pair has same signed correlation coefficients under both conditions. "diff signed" indicates that the gene pair has opposite signed correlation coefficients under two conditions and only one of them passed the criteria that the absolute correlation coefficients greater than rth and q value less than qth. "switched opposites" indicates that the gene pair has opposite signed correlation coefficients under two conditions and both of them passed the criteria that the absolute correlation coefficients greater than rth and q value less than qth.}
}
\description{
This function is used to identify differentially coexpressed links (gene pairs) and enriched genes.
}
\details{
diffcoexp function takes two gene expression matrices or data frames under two conditions as input, calculates gene-gene correlations under two conditions and compare them with Fisher's Z transformation, filter the correlation with the rth and qth and the correlation changes with r.diffth and q.diffth.
}
\examples{
#diffcoexp()
}
\author{
Wenbin Wei
}
\keyword{coexpression}
