% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcoexp.R
\name{diffcoexp}
\alias{diffcoexp}
\title{A diffcoexp Function}
\usage{
diffcoexp(exprs.1, exprs.2, rth = 0.5, qth = 0.1, r.diffth = 0.5,
  q.diffth = 0.1, q.dcgth = 0.1, r.method = c("pearson", "spearman")[1],
  q.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY",
  "fdr")[1])
}
\arguments{
\item{exprs.1}{a data frame or matrix for condition 1, with rows as variables (genes) and columns as samples.}

\item{exprs.2}{a data frame or matrix for condition 2, with rows as variables (genes) and columns as samples.}
}
\description{
This function is used to identify differentially coexpressed links (gene pairs) and enriched genes.
}
\examples{
diffcoexp()
}
\keyword{coexpression}
