\docType{data}
\name{exprs.2}
\alias{exprs.2}
\title{exprs.2}
\format{A matrix with 400 genes and 12 samples.}
\usage{
exprs.2
}
\description{
expression of 400 genes in 14 samples (GSM94988 to GSM95001) of yeast after pulses
0.2 g/l glucose, \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4158}.
}
\keyword{datasets}
